move rotate functions to the geometric_transformations module
